ai_engineering/requirements.txt
langchain-google-genai
langchain-huggingface
langchain-community
langchain-core
langchain-text-splitters
langchain
faiss-cpu
pypdf
sentence-transformers
python-dotenv
